README: use nr_cpus to limit the cpu number for kdump to avoid boot hang
Issue: LIN1024-2562

When kdump is triggered to boot on cpu 3-15 on AXM55xx board, kdump
kernel will hang, this is because the booting CPU will be mapped to
cpu_logical_map[0], whereas this is belong to cluster0(cpu0-3),
on AXM55XX board, clusters[0-4] are isolated to each other, it will mess up
the whole system with this logical map and irqs can not be responded
correctly, then boot kernel will failed.

Use nr_cpus will force kernel use default cpu logical map that clusters
and cpus are well mapped and the hang issue can be avoid.

Signed-off-by: Liwei Song <liwei.song@windriver.com>